Exercises_07_18 Done

+++ src/Chapters/Chapter07/VolkanExercises/SelectionSort.java
@@ -2,14 +2,14 @@ package Chapters.Chapter07.VolkanExercises;
 
 public class SelectionSort {
     public static void main(String[] args) {
-        double [] arr ={1,2,3,4,5,6,8,10,15,356,15,58};
+        int [] arr ={1,2,3,4,5,6,8,10,15,356,15,58};
         selectionSort(arr);
-        for (double d:arr) {
+        for (int d:arr) {
             System.out.print(d+" ");
 
         }
     }
-    public static void selectionSort(double [] list){
+    public static void selectionSort(int [] list){
         for (int i = 0; i < list.length-1 ; i++) {
             double min =list[i];
 
@@ -23,7 +23,7 @@ public class SelectionSort {
             }
             if(minIndex!=i){
                 list[minIndex]=list[i];
-                list[i]=min;
+                list[i]= (int) min;
 
             }
